feat: migrate progress over ctl/rpc buses
