+++ src/bus/rpc.rs
@@ -1,22 +1,19 @@
-use std::fmt::{self, Debug, Display, Formatter};
 use std::str::FromStr;
 use std::time::Duration;
 
 use farcaster_core::{
     blockchain::Blockchain, role::TradeRole, swap::btcxmr::PublicOffer, swap::SwapId,
 };
-
-use amplify::{ToYamlString, Wrapper};
+use amplify::{ToYamlString};
 use internet2::addr::{InetSocketAddr, NodeAddr};
 use internet2::Api;
-use microservices::rpc;
 #[cfg(feature = "serde")]
 use serde_with::{DisplayFromStr, DurationSeconds};
 use strict_encoding::{StrictDecode, StrictEncode};
 use uuid::Uuid;
 
-use crate::bus::ctl::{OfferStatusPair, Outcome};
-use crate::bus::request::{List, Request};
+use crate::bus::ctl::{OfferStatusPair, Progress, Outcome, Failure, OptionDetails};
+use crate::bus::request::{List};
 use crate::cli::OfferSelector;
 
 #[derive(Clone, Debug, Display, From, StrictDecode, StrictEncode, Api)]
@@ -68,9 +65,14 @@ pub enum Rpc {
 
     // Progress functionalities
     // ----------------
+    // Returns a SwapProgress message
     #[api(type = 1003)]
     #[display("read_progress({0})")]
     ReadProgress(SwapId),
+
+    #[api(type = 1005)]
+    #[display(inner)]
+    SwapProgress(SwapProgress),
 
     #[api(type = 1006)]
     #[display("subscribe_progress({0})")]
@@ -83,10 +85,6 @@ pub enum Rpc {
     #[api(type = 1002)]
     #[display(inner)]
     Progress(Progress),
-
-    #[api(type = 1005)]
-    #[display(inner)]
-    SwapProgress(SwapProgress),
 
     #[api(type = 1001)]
     #[display(inner)]
@@ -328,13 +326,6 @@ pub struct CheckpointEntry {
     pub trade_role: TradeRole,
 }
 
-#[derive(Clone, Debug, Display, StrictEncode, StrictDecode)]
-#[display(inner)]
-pub enum Progress {
-    Message(String),
-    StateTransition(String),
-}
-
 #[cfg_attr(feature = "serde", serde_as)]
 #[derive(Clone, PartialEq, Eq, Debug, Display, Default, StrictEncode, StrictDecode)]
 #[cfg_attr(
@@ -345,33 +336,6 @@ pub enum Progress {
 #[display(SwapProgress::to_yaml_string)]
 pub struct SwapProgress {
     pub progress: Vec<ProgressEvent>,
-}
-
-#[derive(Wrapper, Clone, PartialEq, Eq, Debug, From, Default, StrictEncode, StrictDecode)]
-#[cfg_attr(
-    feature = "serde",
-    derive(Serialize, Deserialize),
-    serde(crate = "serde_crate")
-)]
-pub struct OptionDetails(pub Option<String>);
-
-impl Display for OptionDetails {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        match self.as_inner() {
-            None => Ok(()),
-            Some(msg) => f.write_str(msg),
-        }
-    }
-}
-
-impl OptionDetails {
-    pub fn with(s: impl ToString) -> Self {
-        Self(Some(s.to_string()))
-    }
-
-    pub fn new() -> Self {
-        Self(None)
-    }
 }
 
 #[cfg_attr(feature = "serde", serde_as)]
@@ -463,103 +427,6 @@ pub enum OfferStatus {
     InProgress,
     #[display("Ended({0})")]
     Ended(Outcome),
-}
-
-/// Information about server-side failure returned through RPC API
-#[cfg_attr(
-    feature = "serde",
-    derive(Serialize, Deserialize),
-    serde(crate = "serde_crate")
-)]
-#[derive(
-    Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Debug, Display, StrictEncode, StrictDecode,
-)]
-#[display("{info}", alt = "Server returned failure #{code}: {info}")]
-pub struct Failure {
-    /// Failure code
-    pub code: FailureCode,
-
-    /// Detailed information about the failure
-    pub info: String,
-}
-
-#[derive(
-    Copy, Clone, Ord, PartialOrd, Eq, PartialEq, Hash, Debug, Display, StrictEncode, StrictDecode,
-)]
-#[cfg_attr(
-    feature = "serde",
-    derive(Serialize, Deserialize),
-    serde(crate = "serde_crate")
-)]
-#[display(Debug)]
-pub enum FailureCode {
-    /// Catch-all: TODO: Expand
-    Unknown = 0xFFF,
-}
-
-impl From<u16> for FailureCode {
-    fn from(value: u16) -> Self {
-        match value {
-            _ => FailureCode::Unknown,
-        }
-    }
-}
-
-impl From<FailureCode> for u16 {
-    fn from(code: FailureCode) -> Self {
-        code as u16
-    }
-}
-
-impl From<FailureCode> for rpc::FailureCode<FailureCode> {
-    fn from(code: FailureCode) -> Self {
-        rpc::FailureCode::Other(code)
-    }
-}
-
-impl rpc::FailureCodeExt for FailureCode {}
-
-impl From<crate::Error> for Request {
-    fn from(err: crate::Error) -> Self {
-        Request::Rpc(Rpc::Failure(Failure {
-            code: FailureCode::Unknown,
-            info: err.to_string(),
-        }))
-    }
-}
-
-pub trait IntoProgressOrFailure {
-    fn into_progress_or_failure(self) -> Request;
-}
-pub trait IntoSuccessOrFailure {
-    fn into_success_or_failure(self) -> Request;
-}
-
-impl IntoProgressOrFailure for Result<String, crate::Error> {
-    fn into_progress_or_failure(self) -> Request {
-        match self {
-            Ok(val) => Request::Rpc(Rpc::Progress(Progress::Message(val))),
-            Err(err) => Request::from(err),
-        }
-    }
-}
-
-impl IntoSuccessOrFailure for Result<String, crate::Error> {
-    fn into_success_or_failure(self) -> Request {
-        match self {
-            Ok(val) => Request::Rpc(Rpc::Success(OptionDetails::with(val))),
-            Err(err) => Request::from(err),
-        }
-    }
-}
-
-impl IntoSuccessOrFailure for Result<(), crate::Error> {
-    fn into_success_or_failure(self) -> Request {
-        match self {
-            Ok(_) => Request::Rpc(Rpc::Success(OptionDetails::new())),
-            Err(err) => Request::from(err),
-        }
-    }
 }
 
 #[cfg_attr(feature = "serde", serde_as)]
